Move InventoryGUIBehindBlocker files to src/, include/ files
